backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
bcrypt==3.2.2
passlib[bcrypt]>=1.7.4
pydantic
openpyxl
python-multipart
pydantic-settings
email-validator
httpx
fpdf
